+++ packages/vuikit/src/table/__dev__/stories.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/vue'
 storiesOf('Table', module)
   .add('Default', () => require('./table.vue').default)
   .add('Alignment', () => require('./table-alignment.vue').default)
-  .add('Size', () => require('./table-size.vue').default)
+  .add('Dimensions', () => require('./table-dimensions.vue').default)
   .add('Modifiers', () => require('./table-modifiers.vue').default)
   .add('Slots', () => require('./table-slots.vue').default)
 
@@ -13,9 +13,6 @@ storiesOf('Table/Tree', module)
 storiesOf('Table/Columns', module)
   .add('Sort', () => require('./column-sort.vue').default)
   .add('Select', () => require('./column-select.vue').default)
-  .add('Text', () => require('./column-text.vue').default)
-  .add('Date', () => require('./column-date.vue').default)
-  .add('Toggle', () => require('./column-toggle.vue').default)
 
 storiesOf('Table/Filter', module)
   .add('Default', () => require('./filter.vue').default)