+++ packages/vuikit/src/table/__dev__/table-dimensions.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="uk-padding">
 
-    <h1>Sizes</h1>
+    <h1>Dimensions</h1>
 
     <h2>Narrowed</h2>
 
@@ -24,16 +24,15 @@
     <h2>Columns Size</h2>
 
     <vk-table :data="data">
-      <vk-table-column head="Shrinked" cell="name" size="shrinked"/>
-      <vk-table-column head="Expanded" cell="name" size="expanded"/>
-      <vk-table-column head="Shrinked" cell="name" size="shrinked"/>
+      <vk-table-column head="Shrinked" cell="name" width="shrinked"/>
+      <vk-table-column head="Expanded" cell="name" width="expanded"/>
+      <vk-table-column head="Shrinked" cell="name" width="shrinked"/>
     </vk-table>
 
     <vk-table :data="data">
-      <vk-table-column head="Small" cell="name" size="small"/>
-      <vk-table-column head="Medium" cell="name" size="medium"/>
-      <vk-table-column head="Large" cell="name" size="large"/>
-      <vk-table-column-text head="Text" cell="name"/>
+      <vk-table-column head="Small" cell="name" width="small"/>
+      <vk-table-column head="Medium" cell="name" width="medium"/>
+      <vk-table-column head="Large" cell="name" width="large"/>
     </vk-table>
 
   </div>