+++ packages/vuikit/src/table/columns/column-toggle-sort.js
@@ -1,9 +1,0 @@
-import ColumnToggle from './column-toggle'
-import ColumnSort from './column-sort'
-import { assign } from '@vuikit/core/utils/lang'
-
-export default assign({}, ColumnToggle, {
-  name: 'VkTableColumnTextSort',
-  props: assign({}, ColumnSort.props, ColumnToggle.props),
-  headRender: ColumnSort.headRender
-})