+++ packages/vuikit/src/table/columns/column-tree.js
@@ -33,26 +33,25 @@ export default {
   props: assign({}, Column.props),
   render: Column.render,
   headRender: Column.headRender,
-  cellRender (h, { props, data, parent }) {
-    const { $row } = data
+  cellRender (h, { props, data, parent, row }) {
     const { cell, cellClass } = props
-    const cellValue = get($row, cell)
+    const cellValue = get(row, cell)
 
     return h(ElTableTd, {
       props,
       class: cellClass
     }, [
-      Array($row[ROW_LEVEL]).fill(h(TreeIndent)),
+      Array(row[ROW_LEVEL]).fill(h(TreeIndent)),
 
       parent.thereAreSubLevels && h(TreeIndent, [
-        $row[ROW_CHILDREN_COUNT] && h(TreeArrow, {
+        row[ROW_CHILDREN_COUNT] && h(TreeArrow, {
           props: {
-            rotated: parent.isExpanded($row)
+            rotated: parent.isExpanded(row)
           },
           on: {
             click: e => {
               e[ROW_CLICK_PREVENTED] = true
-              parent.toggleExpand($row)
+              parent.toggleExpand(row)
             }
           }
         })